File moved to Utilities -> CallBack.h
