+++ src/shared/CMakeLists.txt
@@ -36,7 +36,6 @@ set(headers
     AuthCodes.h
     ByteBuffer.h
     ByteConverter.h
-    CallBack.h
     CircularQueue.h
     Common.Legacy.h
     CommonDefines.hpp
