+++ src/world/Server/EventMgr.h
@@ -21,8 +21,9 @@
 #ifndef EVENTMGR_H
 #define EVENTMGR_H
 
-#include "CallBack.h"
 #include <map>
+
+#include "Utilities/CallBack.h"
 
 enum EventTypes : uint16_t
 {
